pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-interpretability-toolkit"
version = "0.1.0"
description = "A production-ready toolkit for analyzing and visualizing attention patterns in transformer-based language models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "websockets>=11.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "redis>=4.5.0",
    "wandb>=0.15.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    "scikit-learn>=1.2.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "pre-commit>=3.3.0",
    "ipykernel>=6.23.0",
    "jupyter>=1.0.0",
    "notebook>=6.5.0",
]

tensorflow = [
    "tensorflow>=2.12.0",
]

visualization = [
    "bertviz>=1.4.0",
    "streamlit>=1.28.0",
]

all = [
    "tensorflow>=2.12.0",
    "bertviz>=1.4.0", 
    "streamlit>=1.28.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/llm-interpretability-toolkit"
Documentation = "https://llm-interpretability-toolkit.readthedocs.io"
Repository = "https://github.com/yourusername/llm-interpretability-toolkit"
Issues = "https://github.com/yourusername/llm-interpretability-toolkit/issues"

[project.scripts]
llm-interpret = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]